requirements.txt
pyscreenshot==3.0
Pillow==8.4.0
